Reorganizing OPD test cases in 2 modules, one for base, one for readonly.
